Update patch set 2
Patch Set 2: Code-Review-1

I don't think this is a good approach on running the tests. Instead of doing ssh to the cluster, it's cleaner to have a separate container (that runs on the jumpserver, not in the cluster) that has kubectl in it and can connect to the cluster, but runs locally all the commands.

This way you can put all the necessary tools in the validation container, instead of alterating the cluster.

Patch-set: 2
Label: Code-Review=-1