Add container scripts for k8s conformance test
JIRA: VAL-9

Added the Dockerfile that installs
 - kubectl, e2e.test, ginkgo
 - sonobuoy
 - robot framework

Makefile that can be used to build the container and push it, including
manifest list for both x86_64 and aarch64. Check the README file for
usage.

Until we have the official build in place, test images are available at
cristinapauna/akraino_validation:k8s-latest

Manual testing performed on the images:
- e2e.test command was verified using:
  e2e.test -ginkgo.focus "\[Conformance\]" -ginkgo.noColor -ginkgo.skip \
  'Alpha|Disruptive|Feature|Flaky' -kubeconfig /root/.kube/config -provider local

- sonobuoy was tested with the command:
  sonobuoy run --wait

- robot framework was tested using https://github.com/robotframework/RobotDemo
  robot --name Robot --loglevel DEBUG keyword_driven.robot data_driven.robot \
  gherkin.robot

Change-Id: I9d661895cb85f030993f2205b7ca4115f46ec21b
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>