Update patch set 8
Patch Set 8:

I thought the whole point of using the robot framework is for its result format. If we are allowing to run shell scripts as tests, we are loosing that feature and uniformity, especially as we are allowing the same testcase to be ran in two different ways. 
It seems I don't have access to the jira ticket so I can't see the rationale behind it. Can provide some details on it?

Patch-set: 8
Label: Code-Review=0