Update patch set 9
Patch Set 9:

Yes its good to go with robot for all the testcases but at kubetest is not working with robot so as a work around executing shell script option is provided. Once we have a robot script for kubetest we can fix the not to have shell script option.

Patch-set: 9
Label: Verified=0