Update patch set 10
Patch Set 10:

> Patch Set 9:
> 
> Yes its good to go with robot for all the testcases but at kubetest is not working with robot so as a work around executing shell script option is provided. Once we have a robot script for kubetest we can fix the not to have shell script option.

I feel like I am still missing something. Robot allows you to run shell commands as much as a shell script. As last resort you can invoke the shell script from robot. Who is supposed to provide that robot script that works with kubetest? Are we waiting for something upstream?

My concern is that if we don't keep a consistency we're going to end up with a lot of loose ends and cleanup to do.

Patch-set: 10
Label: Verified=0