Enable running e2e test from a remote server
The approach where kubetest was run by e2e.sh shell script required it
is run in the cluster node, and therefore expected that kubetest was
already installed in the node. Changed to use directly e2e.test for
running the test suite and implemented this as a Robot test case.

Change-Id: Ib26737ef2e5c7af712e17034b9c0233ecaab8ed3
Signed-off-by: Juha Kosonen <juha.kosonen@nokia.com>