Update patch set 3
Patch Set 3:

I realize now that since this test is under security folder, we'll need a security layer container. We can do that in a different patch, but maybe we should add in the jira what tools the container needs. 

I'll try to run the test today and come back with feedback on the patch.

Patch-set: 3