Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> This patch is for building a blueprint specific Docker for testing. I would refer keeping this repo generic and have the specific code in blueprint repos

If we keep the blueprint tests inside the blueprint repo, we will not be able to test the blueprint layer from the validation project. To me, that seems an incomplete level of validation for any given project. And I am also not sure how we're going to interpret the results...

Patch-set: 9