Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I am wondering if helm should be a layer in its own, or should be included in the k8s layer. Juha, what do you think?

I was pondering this too when implementing the patch. Initially was thinking to include in k8s layer but ended up creating an own layer. If it seems more reasonable to have this on k8s layer, that's ok for me. I'm fine with both scenarios (and of course volunteering to do the change).

What's your thinking? Others, any thoughts?

Patch-set: 1