Update patch set 2
Patch Set 2:

I believe we should abandon this patch.

Instead, we should extract the needed information and insert it in the following patch which has the same scope

https://gerrit.akraino.org/r/#/c/ci-management/+/1657/

Currently, the latter triggers the blucon.py for iec and rec. But it does not push results to nexus and it does not support the functionality of notifying UI.

Patch-set: 2