Update patch set 1
Patch Set 1: Code-Review+2

We discussed this internally and concluded that this approach is the way to go. The clusters will upgrade to newer versions of k8s and the intention is to upstream our patches and use the official images going forward. 
So for this release we'll build individually the versions needed for back-porting, and we'll look into dropping building our own kube-conformance image for the next release.

Patch-set: 1
Label: Code-Review=+2