Update patch set 1
Patch Set 1:

That's the approach we're applying for docker_bench: the result of execution is always PASS. I'm not sure if that's a better way either. A user may conclude all is good and no actions required since the test is flagging PASS.

Or we could - whenever there's no clear pass criteria - always flag FAIL. This would be an indication for user that further analysis is required.

As a such, both options are fine by me.

Patch-set: 1